+++ greskell/src/Data/Greskell/Extra.hs
@@ -7,13 +7,14 @@
 --
 -- @since 0.2.3.0
 module Data.Greskell.Extra
-  ( writePropertyKeyValues
+  ( writePropertyKeyValues,
+    writePropertyKeyValues'
   ) where
 
 import Data.Aeson (ToJSON)
 import Data.Greskell.Binder (Binder, newBind)
 import Data.Greskell.Graph
-  ( Property(..), Element
+  ( Property(..), Element, KeyValue(..)
   )
 import qualified Data.Greskell.Graph as Graph
 import Data.Greskell.GTraversal (Walk, SideEffect, gProperty)
@@ -28,6 +29,16 @@ writePropertyKeyValues :: (ToJSON v, Element e) => [(Text, v)] -> Binder (Walk S
 writePropertyKeyValues pairs = fmap mconcat $ mapM toPropStep pairs
   where
     toPropStep (key, value) = fmap (gProperty $ Graph.key key) $ newBind value
+
+-- | Make a series of @.property@ steps to write the given key-value
+-- pairs as properties.
+--
+-- This one allows heterogeneous value types, but it's not based on
+-- 'Binder' monad.
+writePropertyKeyValues' :: Element e => [KeyValue e] -> Walk SideEffect e e
+writePropertyKeyValues' pairs = mconcat $ map toPropStep pairs
+  where
+    toPropStep (KeyValue k v) = gProperty k v
 
 ---- -- | Make a series of @.property@ steps to write all properties in the
 ---- -- given 'PropertyMap'.